Rag_shellokode/backend/requirements.txt
fastapi
uvicorn
python-dotenv
voyageai
qdrant-client
pdfplumber
python-docx
pandas
openpyxl
python-pptx
langchain-text-splitters
streamlit
requests